Added `rCommandsManager`.
